fix: PageOnePage componet renaming
